Fixed unwanted quitting problem
I noticed a problem with the method by which canceling was detected.
Small inputs, especially specifying whether to encrypt or decrypt with
"e" or "d," would cause the program to quit because the algorith was not
concious of words.  It only checked to see if what was typed was
contained within a string of exiting keywords.

This problem has been resolved by prompting the user to type quit and
only accepting this (though being more forgiving of extra input).